# doc/api.yaml
{
  "openapi": "3.0.0",
  "paths": {
...
    "/favs/track/{id}": {
      "post": {
        "operationId": "FavsController_addTrack",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "FavsController_removeTrack",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/track": {
      "post": {
        "operationId": "TrackController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTrackDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Track"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "TrackController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Track"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/track/{id}": {
      "get": {
        "operationId": "TrackController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Track"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "TrackController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTrackDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Track"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "TrackController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
...
  },
  "info": {
    "title": "Home library",
    "description": "The final task of RSSchool NodeJS course 2024Q1",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateTrackDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50
          },
          "duration": {
            "type": "number",
            "minimum": 1,
            "maximum": 1000
          },
          "artistId": {
            "type": "string",
            "nullable": true
          },
          "albumId": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "name",
          "duration",
          "artistId",
          "albumId"
        ]
      },
      "Track": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "duration": {
            "type": "number"
          },
          "artistId": {
            "type": "string",
            "nullable": true
          },
          "albumId": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "duration",
          "artistId",
          "albumId"
        ]
      },
      "UpdateTrackDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50
          },
          "duration": {
            "type": "number",
            "minimum": 1,
            "maximum": 1000
          },
          "artistId": {
            "type": "string",
            "nullable": true
          },
          "albumId": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "name",
          "duration",
          "artistId",
          "albumId"
        ]
      },
...
    }
  }
}